backend/requirements.txt
absl-py==0.13.0
arabic-reshaper==2.1.3
asgiref==3.3.4
astunparse==1.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
clang==5.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-user-agents==0.4.0
djangorestframework==3.12.4
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gcloud==0.17.0
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
gunicorn==20.1.0
h5py==3.1.0
html5lib==1.1
httplib2==0.20.2
idna==2.10
joblib==1.0.1
jws==0.1.3
keras==2.6.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauth2client==3.0.0
oauthlib==3.1.1
opencv-python==4.5.2.54
opt-einsum==3.3.0
phonenumbers==8.12.27
Pillow==8.2.0
protobuf==3.17.3
psycopg2==2.9.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.4.3
PyJWT==1.7.1
pyparsing==3.0.6
PyPDF2==1.26.0
Pyrebase==3.0.27
python-bidi==0.4.2
python-dotenv==0.19.1
python-jwt==2.0.1
pytz==2021.1
PyYAML==6.0
reportlab==3.6.2
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
six==1.15.0
sqlparse==0.4.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==2.2.0
twilio==6.61.0
typing-extensions==3.7.4.3
ua-parser==0.10.0
urllib3==1.26.6
user-agents==2.2.0
waitress==2.0.0
webencodings==0.5.1
Werkzeug==2.0.1
whitenoise==5.3.0
wrapt==1.12.1
xhtml2pdf==0.2.5
